creating the svg element using JavaScript
We first declare constants for the width and height of the -svg- element. Then we use
D3 to select the -body- element and append an -svg- element to it, setting the -width-
and -height- attributes in the process.

We save the result of creating the element into the variable -svg- because we're going
to need it later.