calculating character frequencies
The -input- event is triggered anytime the content of the text area changes, whether from
typing, deleting, pasting, or some other action.

The first thing we do in the event's handler function is initialize a new object for keeping
track of the character frequencies.

This -frequencies- object will use characters for its keys and the number of appearances of that
character for its values.

We then get the -.target- of the event (the text area), get its -.value- (the text), and split it
into its individual characters. -- For each character, we determine the current count for that
character, defaulting to 0 if it hasn't been seen yet.

Then we add 1 to that count and store the new count back in the object.

Once all the characters have been counted, we log the -frequencies- object to the console so we can
check everything is working as expected.

Note that we recalculate the -frequencies- object every time the text changes, rather than just
trying to track added or deleted characters.

This makes it much easier to handle cases where multiple characters are added or removed at once,
for example, when text is pasted into the box.

Load the page, and you should see the text area (the -svg- elements is invisible, but you can see it
in the Elements panel if you want to check it'
is there).