converting the frequency data to an array
First, we use -Object.entries- to convert the -frequencies- object int to an array
of two-element arrays, where the first element is the key and the second element is
the value. We map this array into an array of objects, where the key is stored under
the property -char- and the value is stored under the property -count-.

Next, we want to sort the data by character. -- The -sort- method orders the elements
in an array by applying a comparison function to every pair of elements -a- and -b-,
to determine whether -a- should be sorted after -b- or vice versa.

Here we use the -d3.ascending- comparison function, passing -a.char- and -b.char-,
which means the array will be sorted into ascending alphabetical order based on the
-char- property of each object.