Defined the - update - function
Here we declared an -update- function, which gets called every time the text changes. The function
creates, updates, or deletes the SVG elements needed to render the data, according to the same pattern
we learned earlier for binding data to a selection (using the -data- method) and joining in the
necessary elements with the simple version of the -join- method.

The -join- method returns a selection containing all the current -rect- elements, including
any that were just added. -- Each -rect- is now bound to a single datum, which represents a
single character and the number of occurences of that character.

We set the -width-, -height-, -x-, and -y- attributes as appropriate to create a horizontally
oriented bar chart.

The -width- attribute is set to 5 times the character count, so every new instance of a character
makes the bar 5 pixel wider.

The -height- attribute is a constant value of 10 (all bars are the same height), and the -x- attribute
is a constant value of 20 ( all bars start 20 pixels from the left of the SVG element).

The -y- attribute is set to 20 times the index of the datum, meaning that there will be a bar every
20 pixels, giving 10 pixels of space between each bar.