Creating a scale for the bar widths
First, immediately before the -update- function definition, we create an object describing the margins
of our bar chart diagram. -- These values indicate how far from the edges of the SVG element the
main body of the diagram will be.

Inside the -update- function, we use the -d3.scaleLinear- method to create a scale. This means that
input values map linearly to output values (as opposed to logarithmically, for example).

We set the domain from zero to the max count, using the D3 -max- helper.

The range is set from -margin.left- to -width - margin.right- and give us the position of the right
side of the longest bar.

The -scalelinear- helper gives us a function that maps from the data domain to the visual range,
which we assign to the variable -xScale-

We modify our -width- attribute setting to call the -domain- value -0- and -xScale(d.count)- gives
the horizontal position of the right side of the bar.

To get the width of the bar, we need to substract -xScale(0)- from -xScale(d.count)-, because
the width is just the distance between the left side of the bar and the right side of the bar.

This will give an appropriactely scaled bar width based on the count of each datum and the
maximum count.

We se the -x- attribute of the bar to xScale(0) to enforce the left margin.

Reload the page and start typing into the text area. -- The first time you enter a character, a single
bar will appear, at the maximum width.