scaling the bar heights
To create a scale for the heights fo the bars, we use the d3.scaleBand helper. -- This lets us
create a set of evenly spaced bands. The domain here is slightly different, because instead of
an array giving the minimum and maximum values, it contains the full set of values.

The range here is from -margin.top- to -height -margin.bottom- which gives the range of -y- values
the bars will exist in.

The -padding- value defines how much space there is between bars based of the space available:
 - 0 means that they are as tall as possible and will be touching
 - while 0.5 means that the bars will take up half of the space available

Scales created using -scaleBand- also have a -bandwidth- method that returns the scaled size of the
bands, which we can use to set the height of the bars. -- To get the -y- attribute of the bar,
we pass -d.char- to the -yScale- function, because the domain of this scale is all the characters
present in the data.

Reload the page and type some text in to the text area. The first character you enter will cause
a single tall black bar to appear, but for every unique character you type a new bar will be added,
and the heights of the existing bars will decrease to make space.