adding -g- element for containing axes; and drawing the axes
We create the top axis container by appending a -g- element to the -svg- element and giving it an id
of 'top'.

Because we defined the range for -xScale- to be [margin.left, width - margin.right], that will also
define the visual range of the axis. -- -xScale- doesn't have any knowledge of vertical positioning,
however, which is why we have to translate it down by margin.top

We store the element selection in a variable called - topContainer- so we have a reference to it for
later when drawing the axis into the container.

The left axis container is created similarly, but this time we have to translate it to the right by
-margin.left-, since -yScale- has no knowledge of horizontal positioning.

## Now we can draw the containers

First we call d3.axisTop, passing -xScale- and d3.axisLeft passing -yScale-

This gives us two axis generators, however, we instead pass the generators themselves to a D3 method
called -call-. -- This method, when chained to a selection (like topContainer or leftContainer) --
-- calls the provided function on the current selection.

Thus, writing -topContainer.call(topAxis)-; is equivalent to writing -topAxis(topContainer)-; with
either statement drawing the top axis of the bar graph.

It's considered more idiomatic to use -call- and this makes it easier to chain other methods to the
statement.

Reload the page and type some text in the text area. Should see the axes now, hurray!