rendering integer ticks on the top axis
First we have to get the ticks, which are available using the ticks method on the
-xScale- generator. -- We then filter the ticks to integer values using
 _Number.isInteger_

This will convert an array like [ 0, 0.5, 1, 1.5, 2] -> to [0, 1, 2]

Next, we set the filtered tick values on the top axis using the -tickValues- method

Finally we use the -tickFormat- method to set a rendering format for the numbers.
 - This method takes a formatting function that will be used to format each tick
 value.
 - In this case, d3.format('d') returns a function that formats numbers without
 the decimal point.

Reload the page and enter some tet again; you should see whole numbers rendered
without the decimal point.