updated script.js and style.css
updated the rendering code in script.js to use the -getClass- function
to set a class name on each 'rect' element.

Now every 'rect' element will have a class name based on the character from that
element's datum.

The last step is the CSS that will give each class name a different fill color.

Now when you reload the page and type in some different characters, the bars
should be assigned different colors based on the kind of character enetered.