updated README.md and created a -standardizeSpace- function in script.js
We first declare a -standardizeSpace- function that takes a character and calls the
-trim- method on it. The -trim- method removes whitespace at the start or end of a string.
So if it returns an empty string, we know the character is whitespace. In this case,
we return the string '<space>'.

Otherwise, we return the character unchanged. We then have to modify the text processing
code to call our function and standardize the whitespace characters before using them
as keys in the frequencies object.

Now when you enter various kinds of whitespace characters in the text area, you should see
a single bar labled <space> instead of multiple bars with empty labels.