Animating the bars
First, set a key function to tell D3 that the datum's -char- property
should be used as its identifer.

Next, switch to the advanced join technique. -- The -enter- function first
adds the -rect- element and sets its -x-, -y-, and -class- attributes before
the call to -transition-, meaning that these attributes won't be animated.

The -width- and -height- attributes come after the call to -transition-, so
these attributes WILL be animated.

This way, new elements will grow in place from the left axis.

The -update- function, animates the -width- and -height- again, but also
animates the -y- attribute. -- This means the existing elements will slide up
or down to their new position when new elements are added.

Finally, the -exit- function animates the -width- and -height- to 0 before
the element is removed, causing elements to shrink away to nothing at their
previous position.

Reload the page and try adding and removing characters in the text area. Enjoy
watching how the elements animate in, out, or update.